Major Fix - Remove Driver and Passenger and merge it into User
1. In inheritance, all class is represented as one individual entity
even though it has extended from super class. In our case, User, Driver
and Passenger are considered as different entity and you can't create a
User and consider that User as Driver or Passenger. You need to created
Driver seperately which would have different id and Passenger
seperately. So in nutshell you would end up with 3 users in the system
when you add a user which will perform Driver and Passenger role.