Added equals and hascode method for key domain/data models
1. Added equals and hashcode method for key domain/data models based on
database primary key as well as business keys. Reason for adding this
was to ensure hibernate is able to compare collection items and find out
only incremental item but its still not working and currently it first
delete all the items matching id and then reinsert the difference. Still
investigating but equals and hascode would be useful in future