Major Refactor - One place to define collection type for domain models
1. Currently there were two places where we were defining collection
types one is in domain model and other one is in mapper, which was
overwriding the domain model collection type such as hashset to linked
list while mapping from enity to model. So removed that duplicate code
from mapper and now mapper would use the domain model collection type.
So in nutshell we have one place to define collection type