Major Change - Made most of PK from int to Long. Additional remarks -
1. Bug fixed - For every offered ride, it was matching with only one
ride request irrespective of more seats available. So now this has been
changed and this will match all ride request till all seats are filled
up.

2. TODO later, we are not able to store ride id / ride request id as
long in mongo, as by default it uses Integer and we need to provide more
customization in Json while creating records to store in Long format.
Though this doesn't have any impact immediately, so left it for the time
being.